// src/app/services/vote-service/model/partyVotesDto.ts
/**
 * AllPUDB
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PartyVotesDto { 
    pollingUnitId?: number;
    apc?: number;
    lp?: number;
    nnpp?: number;
    pdp?: number;
    total?: number;
}