// src/app/services/vote-service/model/voteResult.ts
/**
 * AllPUDB
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PartyVotes } from './partyVotes';

export interface VoteResult { 
    id: number;
    name: string;
    item: string;
    partyVotes: PartyVotes;
    percentageComplete: number;
}